AdmobNative/src/test/java/com/hpc/admobnative/WaterfallAdLoaderUnitTest.java
package com.hpc.admobnative;

import org.junit.Test;

import java.util.Random;

import static org.junit.Assert.*;

/**
 * Example local unit test, which will execute on the development machine (host).
 *
 * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
 */
public class WaterfallAdLoaderUnitTest {
    private String[] _unitIds = new String[] {"1", "2", "3"};
    private int _numOfAdsToLoad = 3;

    @Test
    public void success_isCorrect() {
//        final int successIndex = (int)(Math.random() * (_unitIds.length * _numOfAdsToLoad));
//        final int[] loadCnt = {0};
//        IAdUnitLoader testLoader = new IAdUnitLoader() {
//            @Override
//            public void load(String unitId, IAdLoadResult loadResult) {
//                if (loadCnt[0]++ == successIndex) {
//                    loadResult.onSuccess();
//                } else {
//                    loadResult.onFail("test");
//                }
//            }
//        };
//
//        final int[] successCnt = {0};
//        WaterfallAdLoader loader = new WaterfallAdLoader(_unitIds, _numOfAdsToLoad, testLoader);
//        loader.load(new IAdLoadResult() {
//            @Override
//            public void onSuccess() {
//                ++successCnt[0];
//            }
//
//            @Override
//            public void onFail(String errorMsg) {
//                assertTrue("success index: " + successIndex + ", loadCnt: " + loadCnt[0], false);
//            }
//        });
//
//        assertEquals(1, successCnt[0]);
//        assertEquals(((successIndex / _numOfAdsToLoad) + 1) * _numOfAdsToLoad, loadCnt[0]);
        assertTrue(true);
    }

    @Test
    public void fail_isCorrect() {
//        final int[] loadCnt = {0};
//        IAdUnitLoader testLoader = new IAdUnitLoader() {
//            @Override
//            public void load(String unitId, IAdLoadResult loadResult) {
//                loadResult.onFail("test");
//                ++loadCnt[0];
//            }
//        };
//
//        WaterfallAdLoader loader = new WaterfallAdLoader(_unitIds, _numOfAdsToLoad, testLoader);
//        loader.load(new IAdLoadResult() {
//            @Override
//            public void onSuccess() {
//                assertTrue(false);
//            }
//
//            @Override
//            public void onFail(String errorMsg) {
//                assertTrue(errorMsg, true);
//            }
//        });
//
//        assertEquals(_unitIds.length * _numOfAdsToLoad, loadCnt[0]);
        assertTrue(true);
    }
}